PMUFW: rom_interface: Deny power up for unavailable islands
Few islands may not be available on certain device variants which is
indicated by a register in EFUSE. Use this value to return errors in
case of power up attempts for these islands through PM framework.
The protection is added at the lowest level (ROM interface) to ensure
that all paths are covered.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>