PMUFW: pm: Change FSMs of slaves in LPD that have clocks
A lot of slaves use to be modeled as always-on because the PMU did not
control any properties of these slaves (there are no dedicated power
islands and the LPD power parent is always-on). Such slaves are UARTs,
SPIs, CANs, I2Cs, etc.
Now, 2 states of such slaves need to be distinguished - one in which
the slave depends on the PLL that drives its clock and another in which
there is no such a dependency. The PM_CAP_CLOCK property is added to
model the dependency - when the slave does not have PM_CAP_CLOCK flag
set in a state it means that it does not depend on its PLL parent.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>