PMUFW: pm: Added basic clock support in order to model PLL dependencies
The clocks are defined in pm_clock.h/c. The PmClock data structure
contains all information needed to determine the PLL parent and
properly handle the node dependencies toward the PLL states.
All clocks specified in the CRL and CRF modules are modeled.

The PmClockHandle structure is used to model node to clock
dependencies. During the PmClockInitList the lists for clocks/nodes
are created based on statically initialized array of PmClockHandles.
After the initialization, each clock will have a list of nodes that
uses it and each node will have a list of clocks that it uses.
Each PLL has a useCount that counts how many clocks is driven by
that PLL. When a node releases a clock, if the clock has no other
users a PLL that is driving the clock is released as well. Similar
works for the PLL request. Through these request/release calls the
PLL increments/decrements its own useCount. If the useCount of a
PLL drops to zero, the PLL is suspended. Similar is done for
resuming a PLL.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>