PMUFW: pm: Trigger the PmClockRequest/Release on node state change
For power nodes the request is triggered upon entering the on
state, release is triggered after entering the off state.
For slave nodes the request/release for clocks is triggered
when the slave suppose to perform the transition. The clocks for
a slave need to be requested if the slave doesn't require them in
the current state but needs them in the next state. The trigger
is added prior to entering the next state. Otherwise, if the slave
does not use clocks anymore the trigger for clock release is added
after exiting the state in which the clocks were required (note
that a slave may have no clocks, or may have one or multiple clocks)

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Sören Brinkmann <soren.brinkmann@xilinx.com>