PMUFW: pm_power: Fixed bug of not requesting the PLL
Because the power state is first updated and afterwards the
clocks are requested when checking how many nodes currently use
the clock the framework concludes that the clock is already used
(due to that power being on). This is wrong and causes the PLL
to not be requested. The problem is fixed by reversing the order
of operations: first request clocks and let the framework observe
the real node->pll dependencies. After, complete the power state
transition by updating power state to on.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>