PMUFW: pm_clock: Fixed bug in calling PmPllRequest/Release
If the pll source has not changed through PmMmioWrite the PLL shall
not be triggered to change the use count via PmPllRequest/Release.
Due to the lack of the check the PLL use count was wrongly accounted.
The errorneous behavior was triggered by the Linux clock framework
providing the mask for whole clock control register even when only
gate is changed.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Signed-off-by: Filip Drazic <filip.drazic@aggios.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>