gpio: Modified the example to work when LED_bits are configured as output
This patch modifies the xgpio_example to work when LED bits are
configured as output.
Removed reading the state of the data as it always returns zero
when LED bits are configured as output.
Increased the LED_DELAY to make LED blinking more visible.

Signed-off-by: Sai Krishna Potthuri <lakshmis@xilinx.com>
Acked-for-series: Srinivas Goud <sgoud@xilinx.com>