lwip: Detect Xilinx PCS PMA with PHY addr
Older versions of Xilinx PCS PMA IP did not have valid identification
and phy address was a parameter. The latest version has a Xilinx
manufacturer ID and phy address is not exported as a parameter.
Hence, if phy address is passed as zero when Xilinx PCS PMA is present,
scan and detect PHY address based on Xilinx ID.

Signed-off-by: Harini Katakam <harinik@xilinx.com>
Acked-for-series: Punnaiah Choudary Kalluri <punnaia@xilinx.com>