bsp: a53: modified translation table for upper ps DDR
This patch modifies memory attribute for DDR at 0x800000000 -
0xFFFFFFFFF. The range of memory is marked as normal memory for
the size defined in hdf and reserved for the rest of the memory
in that 32GB range. This is done in order to avoid speculative
accesses to the memory which is actually not present but defined
in translation tabel.

Signed-off-by: Kinjal Pravinbhai Patel <patelki@xilinx.com>
Acked-for-series: Anirudha Sarangi <anirudh@xilinx.com>