PMUFW: pm: Clean-up the definitions and use of node types
The node types should be used to know to which type to cast in
order to get the pointer to the structure derived from PmNode.
Removed obsolete node types (if in the future we will need fine-
grained types we will create subtypes).

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>