PMUFW: pm_master: Make available masters configurable
Now, there is an array of all possible masters in the system that
are supported by the PFW and a list of masters available by a
configuration (actually used masters). The list of masters is formed
using the pmMasterHead pointer (static in pm_master.c, always points
to the head of list of masters) and nextMaster pointer added in PmMaster
struct. By adding a master via PmMasterAdd call, the master is linked
in the list of masters. All get functions except PmMasterGetPlaceholder
are changed to search through list of available masters, instead of
array of supported masters. The PmMasterGetPlaceholder is used to get
a master by node ID from the array of possible masters (will be used in
set configuration implementation to get the master struct by node ID).
Initially available masters are APU and RPU_0 (only they are allowed to
load the first configuration object). Therefore, upon the initialization
of the PM add these masters as available.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>