PMUFW: pm: Check for permission to load the object
Initially only APU and RPU_0 are allowed to load the configuration
object. Each subsequent attempt to load the object depends on the
previously set configuration. In set configuration section of the
object are specified the permissions to load the next object. If
the permissions are set to zero, the loading of the configuration
object is locked-down.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>