PMUFW: pm_requirement: Added malloc/free operations for PmRequirements
pmReqData is used as heap and pmReqTop is tracking current top of the
heap (it is the index in pmReqData array). When a slave/master pair needs
to be added the PmRequirementAdd should be called and the new
PmRequirements structure would automatically be allocated, properly
initialized and added into slave's and master's lists of requirements.
The PmRequirementFreeAll should be called prior to loading new
configuration in order to drop the previous configuration and free the
allocated elements from the heap/pmReqData.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>