PMUFW: pm: Track whether the master has set latency requirement using flags
PM_MASTER_SET_LATENCY_REQ flag is set when a master sets the maximum
latency requirement for the slave. If the latency requirements are
not set the latencyReq field is ignored. This is a better approach
compared to the approach of reasoning about whether the latency
requirements are set based on latencyReq value. That also allows the
use of memset for clearing requirements value and avoiding the set of
MAX_LATENCY value to mark that the latency is not set.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>