PMUFW: pm: Added remap address handlers for masters
When the configuration object is set the PMU needs to access provided
address in order to read the content of the configuration object.
In the case of RPU, the provided address may be TCM. The address at
which the PMU sees TCM is not equal to the address provided by the RPU.
That is why the remapping is done.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>