PMUFW: pm: Added PmNodeClass
PmNodeClass contains a collection of nodes of the same type/class and
methods for the common behavior of nodes within the class. Supported
node classes are power, processors and slaves. Instead of having a node
typed ID and then searching to get the function for the nodes type/class,
a better approach is to move ID into PmNodeClass and define a pointer to
the class in each object derived from PmNode. Node classes and their methods
are defined in pm_proc/power/slave.c, what is better compared to the previous
approach (definitions were in pm_node.c) because this way the class specific
behavior doesn't have to be exposed in headers. Currently, the only supported
class method is clearing the configuration of nodes.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>