PMUFW: pm: Restructure and improve getting wake-up latency of a node
Calculating and getting the wake-up latency is specific for each node
class. Therefore, the behavior is removed from pm_node.c, implemented
by each node class and exposed for the use via getWakeUpLatency
method of the node class (function pointer in PmNodeClass structure).
The calculation of wake-up latency for processor nodes was not correct
if the processor is in suspending state and that is fixed.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>