PMUFW: pm: Restructure and improve getting power consumption of a node
Getting power consumtion is specific for each node class, therefore
it is removed from pm_node.c with the aim to be implemented by each node
class. The only function left in pm_node.c is PmNodeGetPowerInfo, which is
node specific. Processor and slave classes use PmNodeGetPowerInfo as their
interface (getPowerData) because they do not have anything special. The
power class needs to implement function for traversing power graph because
the power consumption of a power node is equal to the sum of children's
power consumptions, plus the baseline power consumption of that power node.
Due to the MISRA restrictions, the graph traversal needs to be non-recursive.
This patch introduces PmPowerStack, which is used to implement non-recursive
depth-first search (DFS) of the power graph. As oppose to the common stack-based
search algorithms which push on stack unvisited nodes, implemented algorithm pushes
on stack power parents that need to be processed (visited) later, after their
children are processed. That way, the required stack size is equal to the depth
on power hierarchy (instead of number of nodes to be visited), which is only 2
(power islands + power domains). The stack is used by DFS procedure consisting
of 2 calls: PmPowerDfsBegin which prepares data for the search and
PmPowerDfsGetNext which returns the next node to be processed in DFS fashion.
The DFS implementation will be used for other functions which require DFS power
graph traversal as well.
Getting node's power now properly returns an error if power info for a node is not
provided, compared to the previous solution which returned POWER_MISSING as
the power consumption value.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>