PMUFW: pm: Added PmNodeClass constructors (called once on startup)
There is a node class specific behavior that needs to be executed
once upon the PFW startup. Added method in PmNodeClass to be used
as class constructor. If a node class defines constructor, it will
be called when the power management gets initialized upon the PFW
startup. For processors, constructor disables WFI/WAKE events.
For power domains the constructor hooks supply check handlers (if
a domain has defined the handler). That constructor is called by the
constructor of the PmPower class.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked for series: Will Wong <willw@xilinx.com>