lib: xilpm: Add config data generator
Config data object holds the information on system configuration
done in PCW and exported to HDF. This data is extracted during
xilpm generation and saved as an array in pm_cfg_obj.c, which
can be used in apps like FSBL to make the set_configuration call.
We also add XPAR_XILPM_ENABLED to xparameters.h, which can be used
to enable conditonal code compilation in apps that use this library.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-for-series: Sarat Chand Savitala <saratcha@xilinx.com>