bsp: a53: added support for non-secure el1 for a53 64bit
This patch modifies standalone bsp to add support for hypervisor
guest baremetal el1 non-secure application for cortex-a53 64bit
mode. If hypervisor_guest is selected as True, the bsp will be
built for EL1 Nonsecure and if hypervisor_guest is selected as
false, the bsp will be built for EL3 secure. The default option
is hypervisor_guest as false i.e. EL3 Secure.

Signed-off-by: Kinjal Pravinbhai Patel <patelki@xilinx.com>
Acked-for-series: Anirudha Sarangi <anirudh@xilinx.com>