hdcp1x : Removing polling on the XorInProgress HDCP status bit in enabling encryption.
Removing polling on the XorInProgress bit in enabling hdcp cipher encryption. As this bit is set when the hardware latches on the control signals in the incoming video data, we can depend on the hardware to set this bit in the status register after we have enabled XorInProgress in the control register.

Signed-off-by: Yashovardhan Bhatt <yashova@xilinx.com>
Acked-by: Siva Rajesh Jarugula <sivaraj@xilinx.com>