PMUFW: pm_periph: Unify FPD and LPD slaves into one, generic slave
FPD and LPD slaves were separated because LPD was considered always-on,
and it wasn't even modeled. Thereby, there was no dependency modeled
between an LPD slave and the LPD power node, while FPD slaves did depend
on FPD as power parent. Since LPD can go down as well, from the
perspective of the FSM and power parent dependency, the LPD and FPD slaves
do not differ at all. Accordingly, they are unified as generic slaves
which have basically 2 states: one in which they are not used and they
do not have any dependencies on their clocks or power parent states;
and another state in which they are used so their clocks needs to be
running and their power parent needs to be on. There is nothig else
that PFW power management controls regarding the states of these slaves.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>