PMUFW: pm_usb: Fix USB FSM to properly model power parent dependencies
USB FSM was not updated to properly model dependencies toward power parent
when LPD node was added. In unused state, the USB does not have any
dependencies toward power and clocks. In OFF state, although the USB itself
is powered down it requires LPD parent to be on because of the wake-up
logic. In ON state, USB has dependencies toward both clocks and power parent.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>