PMUFW: pm: Added PmSlaveTcm struct for TCM slaves
TCMs have specific properties that other SRAM nodes (L2$ and OCMs)
do not have, so the PmSlaveTcm is the new struct added to model TCMs.
The struct is inherited from PmSlaveSram. The FSM for TCMs is
similar to the basic SRAM: states and transitions are the same, but
the FSM handler for TCM in addition does the ECC init when TCM
transits from OFF to ON state.
The patch also replaces direct word-based writes with memset in
ECC init handlers.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>