PMUFW: pm: Added PmSlaveClass and PmSlaveTcmClass
Slave class is added to provide methods to be called for nodes derived
from slaves. E. g. initialization of the TCM node cannot be done via
basic PmSlave, because the initialization requires knowledge about the
details not modeled in PmSlave (RPU power island needs to be requested if
TCM is initially ON). The PmSlaveTcmClass is added with the init method
to be called for a TCM when the TCM slave gets initialized in PmSlaveInit.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>