PMUFW: pm: Force down unusable nodes after the configuration is set
If a node cannot be used according to the set configuration the PFW
power management forces the node down. The logic for concluding whether
a node is usable is the following:
- Slave is unusable if no master has permissions to use it
- Processor is unusable if it has no master
- Power is unusable if no master has permission to directly request it
  and all its children are unusable as well
- PLL is unusable if it's use count is zero.

Note, the checking and forcing down unusable node must be performed
after the initialization of nodes.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>