PMUFW: EM: Add XPPU and XMPU error handling
When XPPU/XMPU blocks an access due to a violation, it triggers
an interrupt which is routed to all processors and PMU. Ideally
this interrupt can be handled by any processor on the system, but
if XMPUs are locked up using a lock bit, the ISR register can never
be accessed by any processor and thus restricting the ability to clear
interrupts. Though this does not affect the protection function of
XMPU/XPPU, it prevents us from handling further interrupts related to
these violations. To work-around this limitation, PMU is given exclusive
access (via a default in PCW) to XMPU/XPPU registers (instead of setting
lock-bit) and can handle the interrupts, which are routed as part of
error interrupts.This patch adds the infrastructure required to handle
these interrupts. As a default implementation, notify the error in
debug mode and leave it to the user to customize the response depending
on use-case.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-for-series: Will Wong <willw@xilinx.com>