PMUFW: pm: Configure propagation of wake events
The propagation of external wake event needs to be enabled when the
first master requests so and disabled when the last target master
wakes up, or no master has enabled the wake source anymore (e.g. a
master is woken up by another master and it no longer needs external
wake to be enabled). GIC Proxy events do not have config method
because they're not individually controlled. Configure wake event
properly when a master changes the state.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>