PMUFW: pm: Move getting pointer to a master from ORed IPI mask in the function
We need to reuse the code which extracts one-hot encoded IPI mask from ORed
IPI mask of multiple masters and gets the pointer of the master based on
one-hot encoded IPI mask. Thereby, the code is moved in the
PmMasterGetNextFromIpiMask function in pm_master.h/c.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>