bsp:a53:Fix issue in Xil_DCacheDisable API
The existing Xil_DCacheDisable API first flushes the D caches and then
disables it. The problem with that is, potentially there will be a small
window after the cache flush operation and before the we disable D
caches when we might have valid data in cache lines. In such a scenario
disabling the D cache can lead to unknown behavior.
The ideal solution to this is to use assembly code for the complete API
and avoid any memory accesses. But with that we will end up having a
huge amount on assembly code which is not maintainable.
The patch uses a mix of assembly and C code. Puts all local variables in
registers. It avoids making any function calls to avoid using stacks.
The patch also updates the BSP changelog with this fix.

Signed-off-by: Anirudha Sarangi <anirudh@xilinx.com>
Acked-for-series: Kedareswara rao Appana <appanad@xilinx.com>