lwip: Add jumbo frame support for ZynqMP ethernet
Add support to use jumbo frame sizes for ZynqMP GEM.
This support can be enabled by enabling USE_JUMBO_FRAMES
along with the appropriate MTU and buffer size updates in
lwip configuration options.
Since Zynq GEM does not support jumbo frames, a warning
is added to the tcl when jumbo is selected with this MAC
in place. But it is allowed to accomodate for designs using
Zynq GEM and AXI ethernet.
Internal checks are in place to enable jumbo frame support
only for ZynqMP. The changes include enabling jumbo
in GEM register settins, increasing max MTU and buffer sizes
and updating length masks.

Signed-off-by: Harini Katakam <harinik@xilinx.com>
Acked-for-series: Anirudha Sarangi <anirudh@xilinx.com>