PMUFW: pm: Waking up a processor needs to go via master
Master needs to prepare for waking up a processor, so all wake-up
requests now go first to the master, which does the preparation and
then wakes up the processor. Processor can be woken up using
PmMasterWakeProc only if it has a saved resume address.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>