PMUFW: pm_core: Fix PmSystemShutdown behavior
The shutdown type has only one subtype applicable - it applies only
to the caller itself. If a shutdown needs to cover a wider scope, the
action needs to be communicated among the masters.
For the restart (reset) type, 3 different subtypes are applicable
and all 3 are supported in this patch.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>