PMUFW: pm: Fix the use of prealloc requirements provided in config object
Prealloc (pre) requirements of a master represent requirements which
a master has on cold boot and which it cannot request for itself,
because it will start using them even before it is capable of making
a request (e. g. memories or IPIs used for PM API communication).
These prealloc requirements needs to be properly requested before
a master transitions from killed fo active state.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>