PMUFW: pm: Add support for system-level requirements
Some slaves are required for the system to operate, e. g. OCM which
stores FSBL must preserve context independently of the requests made
by masters. In order to ensure that OCM does not get powered down
when APU is forced down, the system-level requirements are introduced
to model such a dependency. For managing system requirements we use
the same infrastructure as for managing master's requirements, except
that in this case there is no master (master pointer is set to NULL).
Currently system requirements exist only for OCM, and state that the
OCM must preserve the context all the time.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>