PMUFW: pm: Fix handling of system requirements in slave force down case
When a slave is forced down we should not allow entering a state
which breaks the system requirements. Therefore, the system
requirements need to be tracked via PmRequirement structure using
a dedicated flag in info variable. Now, if a slave suppose to be
forced down (put in lowest possible power state), the state that will
be entered is constrained such that it still meets system requirements.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>