PMUFW: pm: Swap the meaning of wake/suspend permissions
The wake/suspend permissions variables in PmMaster use to encode
which other masters the master is allowed to request to wake-up/
suspend. However, the configuration object generator seem to
have assumed the opposite: permissions encode which masters are
allowed to wake/suspend the master. Fix the issue in PFW rather
than in configuration object generator.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-for-series: Will Wong <willw@xilinx.com>