Add special timing mode for BS equal timing
Currently with 1080p, the BS symbol appear every 1000 clks,
but at the vertical blanking start, it will be 1051 cycle,
at the beginning of next frame, it will be 950 cycle.
By reducing the frontporch and hsyncwidth value that we are
feeding into VTC, and increasing the same amount to be the
backporch, we can minimize the swing.

Signed-off-by: Tejas Upadhyay <tejasu@xilinx.com>
Acked-by: Anand Dumbre<anandash@xilinx.com>