PMUFW: pm: pll: Add NULL pointer checks
PllRequest and PllRelease functions need to check for NULL pointers
before using the passed PLL object. This is already causing issues when
PllRelease is called from PmClockSnoop and NULL pointers are passed in.
It was leading to weird prints and corruption of DDR locations.

The function return values are also changed accordingly.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-for-series: Will Wong <willw@xilinx.com>