PMUFW: pm: Change model for RPU power island and resolve TCM dependencies
If RPU island resets are released only when RPU power island gets powered
up, resets are not released in the scenario where RPU is initially powered
up and resets are asserted (what was the case without this patch). The
handling of RPU resets now works as follows: when at least one TCM bank is
in ON state, resets are released. PFW never asserts these resets (this is
unchanged behavior, PMU-ROM asserts resets when powering down RPU island).
Since RPU power island has properties that basic PmPower does not model
(dependencies of TCM banks toward the state of RPU island) the PmPowerIslandRpu
is derived from PmPower to model the difference. Whenever a TCM bank is ON the
RPU power island needs to be ON as well, RPU PGE and AMBA resets must
be released. If RPU processors are in reset they need to be halted and their
resets must be released to allow accessing the TCMs.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>