PMUFW: pm: Added probe for RPU nodes in PmGetNodeStatus to allow state detection
This patch is a workaround because the FSBL does not inform PFW about
the booted processors. The RPU processors can be booted initially by
the FSBL (assumed in PFW) or later by openAMP (coprocessor use case).
In order to ensure that the states of RPU processors are initialized
properly for the openAMP use-case, the probing of RPU states is done
when the PmGetNodeStatus is called for the first time for an RPU node
(NODE_RPU, NODE_RPU_0 or NODE_RPU_1).

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>