xilpm: cfg: Add support for RPU lockstep mode
PCW doesnt have an option to specify if RPU is in lock-step mode
or split mode. This decision is made at the time of generating
boot image. So here, we interpret users intent for the mode based
on the sub-system configuration. If only RPU0 sub-system is created
and RPU1 is not added to any sub-system, then we consider it as
lock-step, all other configs as split. Also TCM assignments are fixed
as per the mode.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>