PMUFW: restart: APU WDT triggered restart feature
This patch adds support for APU sub-system restart when FPD
watchdog timer expires. ENABLE_RECOVERY flag needs to be defined
for this feature to be enabled. The restart is handled by first
attempting to idle APU by sending an IPI-1 to ATF and then escalated
(if ENABLE_ESCALATION is defined ) to a PS-Only or SRST based on if
a PL is configured.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>