PMUFW: pm_slave: Add PmGenericSlaveProbe to determine slave state upon init
Initial states of generic slaves are hardcoded to 'unused'. If a slave
is preallocated according to the set configuration, the assumption
about the initial state may not be correct. The initial state of a
preallocated slave should be determined by ORed requirements of all
masters which use the slave upon init.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>