PMUFW: pm: Distinguish models of PL slave from PLD power domain
NODE_PLD is a power domain, NODE_PL is a slave. The slave node is
added to allow preallocation of PL and to ensure that the PL does
not get powered down if the bitstream is loaded by the FSBL (which
does not request the node). To support that scenario, the NODE_PL
slave should be preallocated to the APU and PFW would not power it
down when the configuration is set.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>