PMUFW: pm_slave: Force down unused slaves with private power islands
If GPU_PP_0/1 or USB_0/1 are not requested (preallocated) when the
configuration is set the PFW will power them down. This means that
the islands will be powered up only upon the request.
This is a workaround for the configuration object containing
permissions for APU even when the APU will never use these nodes.
On the other hand, the nodes need to be powered down, otherwise
they would prevent FPD from being powered down.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>