PMUFW: pm_clock: Handle dependencies of TTC clocks toward LSBUS clock
The select control for TTC clock multiplexers is in IOU_TTC_APB_CLK.
Clock lsbus drives TTCs by default and this patch is needed to ensure
that the PLL which drives lsbus does not get suspended while the TTCs
are running.
However, if the select for a TTC mux is changed, this patch is not
sufficient to handle the dependency. It is recommended to use default
mux setting until this gets resolved.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>