drivers: scugic: consider port width while generating interrupt ids
Lets say xlconcat in ports are three (In0, In1 and In2) and In0 port
width is 2 and In1 is 1. If we are generating id for In1, we need to
consider port width of In0 as 2 not 1. so for this added new api
get_port_width to get the width of port and add that and generate
interrupt.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Acked-for-series: Mubin Sayyed <mubin.usman.sayyed@xilinx.com>