lwip: Update correct compiler details even when no ethernet is found
When no ethernet IP is found in the design, lwip tcl prints an error
and exits. The default Makefile.config, however, has A9 compiler and
endian settings. This is misleading on negative tests or when lwip
is accidentally included on an MB or A53 or A9 design with NO ethernet
MAC. Hence default to the correct compiler and endian setting before
erroring out.

Signed-off-by: Harini Katakam <harinik@xilinx.com>
Acked-for-series: Anirudha Sarangi <anirudh@xilinx.com>