axietherent: fix race condition in the driver for multi mac design
In the multi mac fifo based design case if interrupts
are not connected at the h/w level current driver is not
exporting the AXIENET_CONNECTED_TYPE info to the xparameters.h
file for the 2nd IP instance causing hang when user try to run
application for the 2nd IP instance.

This patch fixes this issue.

Acked-for-series: Anirudha Sarangi <anirudh@xilinx.com>
Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>