bsp: standalone: Generate xparameter to indicate LPD usage of CCI
If any LPD peripheral is configured to use CCI, then PMUFW needs to be aware
and keep the FPD powered up. So generate XPAR_LPD_IS_CACHE_COHERENT
if cache coherency is enabled for any one of the LPD masters. This
parameter generation is limited to psu_pmu processor bsp, since it is
intended to be used only in PMUFW.

Signed-off-by: Jyotheeswar Reddy <jyothee@xilinx.com>
Acked-for-series: Will Wong <willw@xilinx.com>