PMUFW: pm_slave: Always set requirements of masters when slave is configured
Current requirements always reflect the requirements of a master, regardless
of the preallocation or PmRequestNode calls. Initially, it is assumed that
all permissible masters require the slave in its highest state. The first next
time when this may be overwritten is when prealloc section is processed, or
later when a master requests the node. Then, 'info' will be set to remember
that the slave has been requested.
Accordingly, in PmGetMaxCapabilities the info is not checked anymore.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>