PMUFW: pm_requirement: Rename PM_MASTER_USING_SLAVE_MASK macro
Macro is renamed because it doesn't mark that a master is using
a slave. Instead, it labels that the master has requested the slave.
Now, a slave can be used by a master even if it is not requested if
the configuration object specifies that the master has permissions
to use the slave and the master is in uninitialized state.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>