sw_apps:zynqmp_fsbl: Fix to allow disabling BSP optimization for FSBL
Currently BSP properties in SDK GUI allows disabling optimization for BSP by
editing extra_compiler_flags. Subsequently, when BSP is built again, optimization
flags for BSP are again applied as the application for which it is being built
is FSBL. To fix this, one more configurable option is provided in BSP properties
screen in SDK in which we can force disable optimization.

Signed-off-by: Sarat Chand Savitala <saratcha@xilinx.com>
Acked-by: Anirudha Sarangi <anirudh@xilinx.com>