PMUFW:xpfw_user_startup: Enable broadcating of inner shareable transactions
For CCI support in linux, the broadcasting of inner shareable
transcations is need to be enabled. The LPD_SLCR_LPD_APU register
can be configured to do so, only when APU is in reset.
This patch updates XPfw_UserStartUp function to enable the broadcating of
inner shareable transactions, if any of the LPD/FPD peripheral
is configured to use CCI.

Signed-off-by: Mubin Sayyed <mubinusm@xilinx.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>