PMUFW: pm_gpp: Release reset after powering up a GPU pixel processor
The PMU-ROM asserts GPU PP's reset when powering down, but it doesn't
release reset on powering up (which is ok). However, linux driver doesn't
release resets either, but it assumes resets are released, causing the
driver to crash. To solve the issue, the GPU PP's reset is released after
the powering up. This holds for both GPU PP 0/1.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Tested-by: Hyun Kwon <hyun.kwon@xilinx.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>