PMUFW: pm: Restore clocks config when APU reboots regardless of node's state
The issue with clock restoration being done in PmClockRequest is as
follows. When the master is rebooted, it is first forced down and then
all the resources it had preallocated on boot are automatically
requested. If the master is exclusive user of a slave, the slave will
change the state during this process, e. g. a generic slave will become
unused and due to the preallocation it will become used again. When the
slave changes state from unused to used, the clocks are requested. So in
that flow, it was fine to restore the clock configuration when clocks
of the node are requested. However, if a slave is shared at the time
when APU is rebooted, the slave state will not change and the
PmClockRequest will never be called. Consequently, clocks will not be
restored. This patch moves the restoration of clocks to always happen
when the nodes are preallocated due to the reboot.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Signed-off-by: Filip Drazic <filip.drazic@aggios.com>
Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>