PMUFW: pm: Clear power down request bit when processor is forced down
It can happen that a processor gets forced down while it already has
its power down request bit set. If the PMU is not clearing the bit,
the processor may later wrongly conclude that it is doing resume instead
of a cold boot. Moreover, if it executes a WFI prior to clearing the bit
on its own the execution of WFI would propagate to the PMU as a power
down request. This patch adds clearing of the bit when a processor
is forced down.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Jyotheeswar Reddy Mutthareddyvari <jyothee@xilinx.com>